refactor: extract culture preferences into shared HttpContext extension (#206)
